refactoring movie list state to separate file
