+++ lib/ui/widgets/movie_popular_list/movie_popular_list_widget.dart
@@ -99,11 +99,26 @@ class _MoviePopularListRowWidget extends StatelessWidget {
                     crossAxisAlignment: CrossAxisAlignment.start,
                     children: [
                       const SizedBox(height: 20.0),
-                      CustomMovieListTextWidget(text: movie.originalTitle, maxLines: 1, fontSize: null, color: Colors.white, fontWeight: FontWeight.bold),
+                      CustomMovieListTextWidget(
+                          text: movie.originalTitle,
+                          maxLines: 1,
+                          fontSize: null,
+                          color: Colors.white,
+                          fontWeight: FontWeight.bold),
                       const SizedBox(height: 5.0),
-                      CustomMovieListTextWidget(text: movie.releaseDate, maxLines: 1, fontSize: 13, color: AppColors.genresText, fontWeight: null),
+                      CustomMovieListTextWidget(
+                          text: movie.releaseDate,
+                          maxLines: 1,
+                          fontSize: 13,
+                          color: AppColors.genresText,
+                          fontWeight: null),
                       const SizedBox(height: 15.0),
-                      CustomMovieListTextWidget(text: movie.overview ?? '', maxLines: 3, fontSize: 12, color: AppColors.genresText, fontWeight: null),
+                      CustomMovieListTextWidget(
+                          text: movie.overview ?? '',
+                          maxLines: 3,
+                          fontSize: 12,
+                          color: AppColors.genresText,
+                          fontWeight: null),
                     ],
                   ),
                 ),
